requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
anthropic>=0.40.0
python-dotenv==1.0.0
debugpy==1.8.0
redis==5.0.1

